move logic for creating cluster
