+++ cmd/cl001/zz_generated.command.go
@@ -9,7 +9,6 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/giantswarm/awscnfm/v12/cmd/cl001/ac000"
-	"github.com/giantswarm/awscnfm/v12/cmd/cl001/ac001"
 	"github.com/giantswarm/awscnfm/v12/cmd/cl001/ac002"
 	"github.com/giantswarm/awscnfm/v12/cmd/cl001/ac003"
 	"github.com/giantswarm/awscnfm/v12/cmd/cl001/ac004"
@@ -57,20 +56,6 @@ func New(config Config) (*cobra.Command, error) {
 		}
 
 		ac000Cmd, err = ac000.New(c)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-	}
-
-	var ac001Cmd *cobra.Command
-	{
-		c := ac001.Config{
-			Logger: config.Logger,
-			Stderr: config.Stderr,
-			Stdout: config.Stdout,
-		}
-
-		ac001Cmd, err = ac001.New(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -263,7 +248,6 @@ func New(config Config) (*cobra.Command, error) {
 	f.Init(c)
 
 	c.AddCommand(ac000Cmd)
-	c.AddCommand(ac001Cmd)
 	c.AddCommand(ac002Cmd)
 	c.AddCommand(ac003Cmd)
 	c.AddCommand(ac004Cmd)
