--- pkg/action/cl001/ac001/explainer.go
@@ -1,11 +1,0 @@
-package ac001
-
-import (
-	"context"
-)
-
-func (e *Explainer) explain(ctx context.Context) (string, error) {
-	s := "Create a basic Tenant Cluster with all its defaults."
-
-	return s, nil
-}
